[ADAM-848] TwoBitFile now support nBlocks and maskBlocks
Fixes #848.
